requirements.txt
streamlit
streamlit-webrtc
openai
plotly
pandas
pillow
